refactor: 메인페이지 진료부분 컴포넌트화 완료
