+++ wanchcoach/src/app/(route)/mainpage/treatment/calendar/[id]/page.tsx
@@ -9,7 +9,7 @@ import {
   TreatmentCalendarItems,
   TreatmentItems,
 } from "@/app/util/controller/treatmentController";
-import TreatmentPaperCalendar from "@/app/_components/Mainpage/TreatmentPaperCalendar";
+import TreatmentPaperCalendar from "@/app/_components/Mainpage/Treatment/TreatmentPaperCalendar";
 
 export default function Calendar({ params }: { params: { id: number } }) {
   const familyId = params.id;
@@ -25,52 +25,43 @@ export default function Calendar({ params }: { params: { id: number } }) {
   };
 
   useEffect(() => {
-    /* 달 별 복약 데이터 api 호출 */
     const fetchData = async () => {
-      try {
-        let response;
-        if (familyId == 0) {
-          response = await TreatmentCalendarForAllController({
-            year: selectedDate.year(),
-            month: selectedDate.month() + 1,
+      if (familyId == 0) {
+        TreatmentCalendarForAllController({
+          year: selectedDate.year(),
+          month: selectedDate.month() + 1,
+        })
+          .then((response) => {
+            setCalendartData(response.data.treatmentDateItems);
+            const dates = response.data.treatmentDateItems.map((item: TreatmentItems) => item.date);
+            setHighlightedDays(dates);
+          })
+          .catch((e) => {
+            console.log(e.message);
+            return;
           });
-        } else {
-          const data = {
-            familyId: familyId,
-            year: selectedDate.year(),
-            month: selectedDate.month() + 1,
-          };
-          response = await TreatmentCalendarController(data);
-        }
-        setCalendartData(response.data.treatmentDateItems);
-        const dates = response.data.treatmentDateItems.map((item: TreatmentItems) => item.date);
-        setHighlightedDays(dates);
-        console.log("달력 데이터 가져오기 성공:", response);
-      } catch (error) {
-        console.error("데이터 가져오기 실패:", error);
-        // 오류 처리
+      } else {
+        TreatmentCalendarController({
+          familyId: familyId,
+          year: selectedDate.year(),
+          month: selectedDate.month() + 1,
+        })
+          .then((response) => {
+            setCalendartData(response.data.treatmentDateItems);
+            const dates = response.data.treatmentDateItems.map((item: TreatmentItems) => item.date);
+            setHighlightedDays(dates);
+          })
+          .catch((e) => {
+            console.log(e.message);
+            return;
+          });
       }
     };
     fetchData();
+
+    fetchData();
   }, [selectedDate.year(), selectedDate.month()]);
 
-  const extractFirstFamilyColor = (
-    calendarData: TreatmentCalendarItems[],
-    familyId: number
-  ): string | null => {
-    if (familyId == 0) {
-      return "#757575";
-    }
-
-    for (const item of calendarData) {
-      for (const treatmentItem of item.treatmentItems) {
-        if (treatmentItem.familyColor) {
-          return treatmentItem.familyColor;
-        }
-      }
-    }
-    return null;
-  };
   return (
     <>
       <div className={styles.body_container}>
