+++ wanchcoach/src/app/(route)/mainpage/treatment/diagnosis/[id]/diagnosis.module.css
@@ -9,7 +9,6 @@
 .body_container {
   margin: 0 8%;
 }
-
 .body_text {
   font-size: 21px;
   font-weight: 600;
