+++ wanchcoach/src/app/(route)/mainpage/treatment/hospital/[id]/hospital.module.css
@@ -8,7 +8,6 @@
 .body_container {
   margin: 0 8%;
 }
-
 .hospital_list_container {
   height: calc(100vh - 60px - 70px - 60px - 12vh - 45px);
   /* - 바텀바 - 헤더 - 프로필 컨테이너 - 메뉴 컨테이너 - 검색바*/
