+++ wanchcoach/src/app/_components/Mainpage/Treatment/TreatmentPaperCalendar.tsx
@@ -1,48 +1,19 @@
 "use client";
 
-import styles from "./Component.module.css";
-import { SetStateAction, useState } from "react";
+import styles from "./Treatment.module.css";
+import { useState } from "react";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { DateCalendar, DateCalendarProps } from "@mui/x-date-pickers/DateCalendar";
-import dayjs, { Dayjs } from "dayjs";
-import { PickersDay, PickersDayProps } from "@mui/x-date-pickers/PickersDay";
-import Badge from "@mui/material/Badge";
+import { Dayjs } from "dayjs";
 import "dayjs/locale/ko"; // 이거 선언 해야 한글형식 적용 됨!
-import BottomSheet from "./BottomSheet";
-import { DrugDayCalendarData } from "@/app/util/controller/medicationController";
-import DayMenu from "./Medication/DayMenu";
-import MedicationInfoBox from "@/app/(route)/mainpage/medication/taking/[id]/_components/MedicationInfoBox";
-import {
-  TreatmentCalendar,
-  TreatmentCalendarItems,
-} from "@/app/util/controller/treatmentController";
-import TreatmentBox from "@/app/(route)/mainpage/treatment/_components/TreatmentBox";
+import BottomSheet from "../BottomSheet";
+import { TreatmentCalendarItems } from "@/app/util/controller/treatmentController";
+import TreatmentBox from "./TreatmentBox";
+import ServerDay from "./children/ServerDay";
+import { extractFirstFamilyColor } from "./children/Function";
 
-interface ServerDayProps extends PickersDayProps<Dayjs> {
-  highlightedDays?: string[];
-  color?: string;
-}
-
-function ServerDay(props: ServerDayProps) {
-  const { highlightedDays = [], day, outsideCurrentMonth, color, ...other } = props;
-
-  const isSelected =
-    !props.outsideCurrentMonth && highlightedDays.includes(day.format("YYYY-MM-DD"));
-  return (
-    <Badge
-      key={props.day.toString()}
-      overlap="circular"
-      badgeContent={isSelected ? " " : undefined}
-      sx={{ "& .MuiBadge-overlapCircular": { backgroundColor: color } }}
-      variant={isSelected ? "dot" : undefined}
-    >
-      <PickersDay {...other} outsideCurrentMonth={outsideCurrentMonth} day={day} />
-    </Badge>
-  );
-}
-
-interface PaperCalendarProps {
+interface TreatmentPaperCalendarProps {
   familyId: number;
   selectedDate: Dayjs | null;
   handleSelectedDateChange: (newDate: Dayjs) => void;
@@ -58,9 +29,9 @@ export default function TreatmentPaperCalendar({
   highlightedDays,
   handleMonthChange,
   calendarData,
-}: PaperCalendarProps) {
-  const currentDate = dayjs();
+}: TreatmentPaperCalendarProps) {
   const [open, setOpen] = useState(false);
+  const familyColors = extractFirstFamilyColor(calendarData, familyId);
   const handleDateChange: DateCalendarProps<Dayjs>["onChange"] = (newDate) => {
     if (!newDate) return;
 
@@ -82,24 +53,6 @@ export default function TreatmentPaperCalendar({
     setOpen(!open);
   };
 
-  const extractFirstFamilyColor = (
-    calendarData: TreatmentCalendarItems[],
-    familyId: number
-  ): string | null => {
-    if (familyId == 0) {
-      return "#757575";
-    }
-
-    for (const item of calendarData) {
-      for (const treatmentItem of item.treatmentItems) {
-        if (treatmentItem.familyColor) {
-          return treatmentItem.familyColor;
-        }
-      }
-    }
-    return null;
-  };
-  const familyColors = extractFirstFamilyColor(calendarData, familyId);
   return (
     <>
       <div>
