+++ wanchcoach/src/app/_components/Mainpage/Treatment/children/Function.tsx
@@ -1,0 +1,19 @@
+import { TreatmentCalendarItems } from "@/app/util/controller/treatmentController";
+
+export const extractFirstFamilyColor = (
+  calendarData: TreatmentCalendarItems[],
+  familyId: number
+): string | null => {
+  if (familyId == 0) {
+    return "#757575";
+  }
+
+  for (const item of calendarData) {
+    for (const treatmentItem of item.treatmentItems) {
+      if (treatmentItem.familyColor) {
+        return treatmentItem.familyColor;
+      }
+    }
+  }
+  return null;
+};
