+++ wanchcoach/src/app/_components/Mainpage/Treatment/children/ServerDay.tsx
@@ -1,0 +1,27 @@
+import { PickersDay, PickersDayProps } from "@mui/x-date-pickers/PickersDay";
+import Badge from "@mui/material/Badge";
+import "dayjs/locale/ko";
+import { Dayjs } from "dayjs";
+
+interface ServerDayProps extends PickersDayProps<Dayjs> {
+  highlightedDays?: string[];
+  color?: string;
+}
+
+export default function ServerDay(props: ServerDayProps) {
+  const { highlightedDays = [], day, outsideCurrentMonth, color, ...other } = props;
+
+  const isSelected =
+    !props.outsideCurrentMonth && highlightedDays.includes(day.format("YYYY-MM-DD"));
+  return (
+    <Badge
+      key={props.day.toString()}
+      overlap="circular"
+      badgeContent={isSelected ? " " : undefined}
+      sx={{ "& .MuiBadge-overlapCircular": { backgroundColor: color } }}
+      variant={isSelected ? "dot" : undefined}
+    >
+      <PickersDay {...other} outsideCurrentMonth={outsideCurrentMonth} day={day} />
+    </Badge>
+  );
+}
