+++ wanchcoach/src/app/util/format/dateFormat.tsx
@@ -52,3 +52,11 @@ export const calculateDDay = (targetDate: string) => {
     return <div className={styles.treatment_box_notdday}>{`D - ${dDay}`}</div>;
   }
 };
+
+export const getCurrentDate = () => {
+  const today = new Date();
+  const year = today.getFullYear();
+  const month = `${today.getMonth() + 1}`.padStart(2, "0"); // month is zero-indexed
+  const day = `${today.getDate()}`.padStart(2, "0");
+  return `${year}-${month}-${day}`;
+};
